Create new package models in data package. The models package separate to sub packages local models and server models. As it sounds, the server models package should contain only models that related to the server. And the local models package should contain only models that related to db.
