+++ app/src/main/java/com/example/composetemplate/data/remote/responses/ExampleResponse.kt
@@ -1,5 +1,6 @@
 package com.example.composetemplate.data.remote.responses
 
+import com.example.composetemplate.data.models.server_models.Source
 import kotlinx.serialization.Serializable
 
 /** Here is an example of a base server response.
