+++ app/src/main/java/com/example/composetemplate/presentation/screens/main_screens/MainViewModel.kt
@@ -2,7 +2,7 @@ package com.example.composetemplate.presentation.screens.main_screens
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.composetemplate.data.local.db_models.Test
+import com.example.composetemplate.data.models.local_models.Test
 import com.example.composetemplate.repositories.TestRepository
 import com.example.composetemplate.utils.LogsManager
 import com.example.composetemplate.utils.tag
